% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_script_file}
\alias{current_script_file}
\title{current script file (in full path)}
\usage{
current_script_file()
}
\value{
The path of the currently executing script or code snippet.
}
\description{
current script file (in full path)
}
\note{
Works with Rscript, source() or in RStudio Run selection
}
