% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customProDB.R
\docType{package}
\name{customProDB}
\alias{customProDB}
\alias{customProDB-package}
\title{CustomProDB}
\description{
Database search is the most widely used approach for peptide and protein identification in
mass-spectrometry-based proteomics studies. Sample-specific protein databases derived from RNA-Seq data can 
better approximate the real protein pools in the samples and thus improve protein identification. 
More importantly, single nucleotide variations, short insertion and deletions and novel junctions 
identified from RNA-Seq data make protein databases more complete and sample-specific. CustomProDB 
enables the easy generation of customized databases from RNA-Seq data for proteomics search. It bridges 
genomics and proteomics studies and facilitates cross-omics data integration.
}
