% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRun.R
\name{easyRun}
\alias{easyRun}
\title{An integrated function to generate customized protein databases for a single sample}
\usage{
easyRun(bamFile, RPKM = NULL, vcfFile, annotation_path, outfile_path,
  outfile_name, rpkm_cutoff = 1, INDEL = FALSE, lablersid = FALSE,
  COSMIC = FALSE, nov_junction = FALSE, bedFile = NULL, genome = NULL,
  ...)
}
\arguments{
\item{bamFile}{Input BAM file name for calculating transcript FPKMs.}

\item{RPKM}{Alternative to bamFile. If non-NULL, must be a vector containing expression level for proteins (e.g. FPKMs from cufflinks).}

\item{vcfFile}{Input VCF file name.}

\item{annotation_path}{The path to read annotation files from (e.g. ids.RData, exon_anno.RData, etc.).}

\item{outfile_path}{Folder path for the output FASTA files.}

\item{outfile_name}{Output FASTA file prefix.}

\item{rpkm_cutoff}{The cutoff for RPKM values. See 'cutoff' in function Outputproseq for more detail.}

\item{INDEL}{Set to TRUE to create a FASTA of the variations due to short insertions and deletions in the VCF (default is FALSE).}

\item{lablersid}{Set to TRUE to include dbSNP rsid as a prefix to each SNP in the FASTA headers (default is FALSE).
If TRUE, the annotation folder must have the dbsnpinCoding.RData file.}

\item{COSMIC}{Set to TRUE to include COSMIC ids in the variation table (default is FALSE).
If TRUE, the annotation folder must have the cosmic.RData file.}

\item{nov_junction}{Set to TRUE to create a FASTA of the peptides that cover novel junctions.
If TRUE, the annotation folder must have the splicemax.RData file, and the bedFile and genome 
parameters must be provided.}

\item{bedFile}{Path to a .bed file which contains the splice junctions identified in RNA-Seq (e.g. junctions.bed from TopHat).}

\item{genome}{A BSgenome object (e.g. BSgenome.Hsapiens.UCSC.hg19::Hsapiens). Default is NULL.}

\item{...}{Additional arguments}
}
\value{
NULL. It creates several FASTA files in <outfile_path>; always creates <outfile_name>_rpkm.fasta
  and <outfile_name>_snv.fasta. Optionally creates <outfile_name>_indel.fasta and <outfile_name>_junc.fasta.
}
\description{
Generate a customized protein database for a single sample.
}
\details{
This function provides a convenient way to generate customized databases for a single sample based on transcript
expression, SNV and INDEL variants, and novel junctions detected during RNA-Seq mapping.
}
\examples{

bamFile <- system.file("extdata/bams", "test1_sort.bam",
            package="customProDB")
vcffile <- system.file("extdata/vcfs", "test1.vcf", package="customProDB")
annotation_path <- system.file("extdata/refseq", package="customProDB")
outfile_path <- tempdir()
outfile_name <- 'test'

easyRun(bamFile, RPKM=NULL, vcffile, annotation_path, outfile_path,
        outfile_name, rpkm_cutoff=1, INDEL=TRUE, lablersid=TRUE,
        COSMIC=TRUE, nov_junction=FALSE)


}
\author{
Xiaojing Wang
}
