% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastComplement}
\alias{fastComplement}
\title{Pure R function to turn a character vector of nucleotides into a reverse complement vector.
When running on many strings, this is much faster than Biostrings::reverseComplement(DNAString).}
\usage{
fastComplement(naString)
}
\arguments{
\item{naString}{a character vector of nucleotides; IUPAC ambiguous nucleotides are supported}
}
\value{
the reverse complement of naString
}
\description{
Pure R function to turn a character vector of nucleotides into a reverse complement vector.
When running on many strings, this is much faster than Biostrings::reverseComplement(DNAString).
}
\examples{
fastComplement(c("GAT", "TACA")) # "CTA"  "ATGT"
fastComplement("MAW") # "KTW"
}
