% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCodingFastaUrlFromUCSC}
\alias{getCodingFastaUrlFromUCSC}
\title{Get URL to download coding sequence FASTA from UCSC genome browser for a given dbkey.}
\usage{
getCodingFastaUrlFromUCSC(dbkey)
}
\arguments{
\item{dbkey}{The UCSC dbkey to get coding sequences for, e.g. hg19, hg38, mm10.}
}
\value{
A URL which can be downloaded with \code{\link{download.file}}
}
\description{
Get URL to download coding sequence FASTA from UCSC genome browser for a given dbkey.
}
\examples{
getCodingFastaUrlFromUCSC("hg38")
}
