% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProteinFastaUrlFromUCSC}
\alias{getProteinFastaUrlFromUCSC}
\title{Get URL to download protein sequence FASTA from UCSC genome browser for a given dbkey.}
\usage{
getProteinFastaUrlFromUCSC(dbkey)
}
\arguments{
\item{dbkey}{The UCSC dbkey to get protein sequences for, e.g. hg19, hg38, mm10.}
}
\value{
A URL which can be downloaded with \code{\link{download.file}}
}
\description{
Get URL to download protein sequence FASTA from UCSC genome browser for a given dbkey.
}
\examples{
getProteinFastaUrlFromUCSC("hg38")
}
