% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantAnnotation.R
\name{getVariantAnnotation}
\alias{getVariantAnnotation}
\title{Get genome annotations for variants called in one or more VCF files.}
\usage{
getVariantAnnotation(vcfFilepaths, ids, exon_anno, proteinseq, procodingseq,
  dbsnpinCoding = NULL, cosmic = NULL)
}
\arguments{
\item{vcfFilepaths}{one or more VCF files to create variant proteins}

\item{ids}{a dataframe mapping between gene/transcript/protein ids}

\item{exon_anno}{a dataframe of exon annotations for each protein coding transcript}

\item{proteinseq}{a dataframe of reference amino acid sequences for each protein}

\item{procodingseq}{a dataframe of reference coding sequences for each protein coding transcript}

\item{dbsnpinCoding}{a GRanges object of dbSNP variants to include in the output tables; default is NULL}

\item{cosmic}{a GRanges object of COSMIC variants to include in the output tables; default is NULL}
}
\value{
A list of annotation tables with named members:
\describe{
\item{variantTable}{describes the SNP variants and their coding consequences}
\item{snvprocoding}{the CDS sequences for the variant proteins from SNPs}
\item{snvproseq}{the translated sequences for the variant proteins from SNPs}
\item{indelprocoding}{the CDS sequences for the variant proteins from INDELs}
\item{indelproseq}{the translated sequences for the variant proteins from INDELs}
}
}
\description{
Get genome annotations for variants called in one or more VCF files. Each sample is called separately
so that variants on the same codon are handled properly. This is a convenience function which calls
OutputVarprocodingseq, OutputVarproseq, Outputaberrant, and OutputNovelJun.
}
\examples{
load(system.file("extdata/refseq", "exon_anno.RData", package="customProDB"))
load(system.file("extdata/refseq", "ids.RData", package="customProDB"))
load(system.file("extdata/refseq", "procodingseq.RData", package="customProDB"))
load(system.file("extdata/refseq", "proseq.RData", package="customProDB"))
load(system.file("extdata/refseq", "dbsnpinCoding.RData", package="customProDB"))
load(system.file("extdata/refseq", "cosmic.RData", package="customProDB"))
vcfFilepath = system.file("extdata/vcfs", "test1.vcf", package="customProDB")

variantAnnotation = getVariantAnnotation(vcfFilepath, ids, exon,
                                         proteinseq, procodingseq,
                                         dbsnpinCoding, cosmic)
head(variantAnnotation$variantTable)
head(variantAnnotation$snvprocoding)
head(variantAnnotation$snvproseq)
head(variantAnnotation$indelprocoding)
head(variantAnnotation$indelproseq)
}
