% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareAnnotationGCF.R
\name{PrepareAnnotationGCF}
\alias{PrepareAnnotationGCF}
\title{Prepare an annotation set from an NCBI GCF assembly.}
\usage{
PrepareAnnotationGCF(ncbiFtpUrl, annotation_path, ...)
}
\arguments{
\item{ncbiFtpUrl}{the location of _cds_from_genomic.fna, _translated_from_cds.faa, and _genomic.gff files}

\item{annotation_path}{specify a folder to store all the annotations.}

\item{...}{additional arguments}
}
\value{
No return value. It creates several .RData files in annotation_path containing annotations needed for further analysis.
}
\description{
Prepare an annotation set from an NCBI GCF assembly.
}
\examples{
annotation_path <- tempdir()
PrepareAnnotationGCF("ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/001/922/835/GCF_001922835.1_NIST_Tur_tru_v1/", annotation_path)

}
\author{
Matt Chambers
}
