% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_or_update_local_cache}
\alias{read_or_update_local_cache}
\title{Read the cached result of an expression from a locally cached file if it exists,
else evaluate the expression, cache it, and return result.}
\usage{
read_or_update_local_cache(expression, local_cache_path, object_name)
}
\arguments{
\item{expression}{The expression to be evaluated and returned if the cached result does not exist.}

\item{local_cache_path}{The path that stores the cached objects. If NULL, no caching is performed.}

\item{object_name}{The name of the object used to read/create the cache file.}
}
\value{
The result of evaluating the expression, either read from cache or from actually evaluating it.
}
\description{
Read the cached result of an expression from a locally cached file if it exists,
else evaluate the expression, cache it, and return result.
}
\note{
If evaluating the expression has side-effects,
they will not happen if the result is returned from cache.
}
\examples{
\dontrun{
read_or_update_local_cache(2+2, tempdir(), "TwoPlusTwo")
}
}
