% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantAnnotation.R
\name{variantType}
\alias{variantType}
\title{Get the type of variant given one or more reference alleles and the corresponding variant alleles.}
\usage{
variantType(ref, alt)
}
\arguments{
\item{ref}{the reference alleles}

\item{alt}{the alternate (variant) alleles}
}
\value{
variant type for each ref/alt pair, one of:
\enumerate{
\item snp (single nucleotide polymorphism)
\item mnp (multiple nucleotide polymorphism)
\item ins (insertion)
\item del (deletion)
\item mix (some combination of the above)
}
}
\description{
Get the type of variant given one or more reference alleles and the corresponding variant alleles.
}
\examples{
ref = c("A", "A",   "AA", "AT", "A",  "AA",  "AA",  "AA",  "ATTA", "AAA",  "A",    "TAAA")
alt = c("C", "C,G", "CT", "A",  "AT", "C",  "CAT", "TAAC", "ATA",  "AATA", "T,AT", "TAA,TAAAA")
print(cbind(ref, alt, variantType(ref, alt)))
}
